# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Photo Caption Contest",
    "version": "1.0.0.0"
  },
  "paths": {
...
    "/api/image/{id}": {
      "get": {
        "summary": "get single image",
        "description": "to retrieve a specific image with an Id from the DB ",
        "operationId": "get_image_id",
        "responses": {
          "200": {
            "description": "to retrive single image from DB",
            "content": {
              "application/json": {
                "examples": {
                  "single image": {
                    "value": {
                      "status": "image retrived successfully",
                      "image": {
                        "id": 37,
                        "caption": "Ayser Shaqruni",
                        "imageUrl": "/uploads/images/1768493850656-912754963.jpg",
                        "captionByUserId": 39,
                        "createdAt": "2026-01-15T16:17:30.669Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the image.",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      },
      "delete": {
        "summary": "to delete an image ",
        "description": "to delete an image  specified by an id ",
        "operationId": "delete_image",
        "responses": {
          "default": {
            "description": "Default error sample response",
            "content": {
              "application/json": {
                "examples": {
                  "deleting an image ": {
                    "value": "{\n    \"status\": \"image deleted successfully\",\n    \"image\": {\n        \"id\": 9,\n        \"imageUrl\": \"/uploads/1768396570013-747098886.jpg\",\n        \"createdAt\": \"2026-01-14T13:16:10.030Z\",\n        \"updatedAt\": \"2026-01-15T08:27:20.394Z\"\n    }\n}"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the image.",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
...
    "/api/user/{id}": {
      "summary": "to rerieve a single user",
      "get": {
        "summary": "to retrieve a single user from the DB specified by id",
        "description": "to get a single user from DB",
        "operationId": "get_user_id",
        "responses": {
          "200": {
            "description": "to get a single user from DB",
            "content": {
              "application/json": {
                "examples": {
                  "single user retrieval ": {
                    "value": {
                      "staus": "success,",
                      "user": {
                        "id": "39,",
                        "username": "Ayser,",
                        "email": "ayser.shaaqruni@yahoo.com"
                      },
                      "passwordHash": "$2b$10$h08QEAN5tWcmk1/DKrT64uNo1w.z6HPXN0nOuxnaCBC8/AcxeZs1m",
                      "createdAt": "2026-01-15T11:28:15.395Z",
                      "updatedAt": "2026-01-15T11:28:15.395Z"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error sample response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      },
      "delete": {
        "summary": "to delete a user with spicified id from the DB ",
        "description": "",
        "operationId": "delete_user",
        "responses": {
          "200": {
            "description": "to delete a single user from DB",
            "content": {
              "application/json": {
                "examples": {
                  "deleting user ": {
                    "value": {
                      "status": "deleted successfuuly",
                      "user": {
                        "id": 25,
                        "username": "ahmed_shaqruni",
                        "email": "ahmed.1983@gmail.com",
                        "passwordHash": "$2b$10$QB3xcm6kGZXORwI5Odigb.srATuEnLZLJcHAopn5Lm0pDdYnFSc7K",
                        "createdAt": "2026-01-15T09:21:39.580Z",
                        "updatedAt": "2026-01-15T09:21:39.580Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error sample response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user to be deleted.",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
...
  },
...
}